third_party/WebKit/LayoutTests/device_orientation/motion/add-listener-from-callback.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/device-orientation-helpers.js"></script>
<script>
'use strict';

async_test(test => {
  assertTestRunner();
  var motionData = generateMotionData(1.1, 2.1, 3.1,
                                      1.2, 2.2, 3.2,
                                      1.3, 2.3, 3.3,
                                      100);

  var firstListenerEvents = 0;
  var firstListener = test.step_func(event => {
    checkMotion(event, motionData);
    window.removeEventListener('devicemotion', firstListener);
    if (++firstListenerEvents == 1)
      window.addEventListener('devicemotion', secondListener);
  });

  var secondListenerEvents = 0;
  var secondListener = test.step_func(event => {
    checkMotion(event, motionData);
    ++secondListenerEvents;
    assert_equals(firstListenerEvents, 1, "Too many events fired for the first listener");
    assert_equals(secondListenerEvents, 1, "Too many events fired for the second listener");
    test.done();
  });

  setMockMotion(motionData);
  window.addEventListener('devicemotion', firstListener);
}, 'Tests that adding a new devicemotion event listener from a callback works as expected.');
</script>
</body>
</html>
